api/requirements.txt
flask_sqlalchemy
flask
flask_cors
tqdm